mitosheet: create more tests
