+++ mitosheet/mitosheet/tests/test_mito_backend.py
@@ -191,48 +191,4 @@ def test_create_backend_with_code_options_works():
     mito_backend = MitoBackend(code_options=code_options)
     assert mito_backend.steps_manager.code_options == code_options
 
-def test_create_backend_with_column_definitions_works():
-    column_definitions= [
-        {
-            'columns': ['A', 'B'],
-            'conditional_formats': [{
-                'filters': [{'condition': 'greater', 'value': 5}], 
-                'font_color': '#c30010', 
-                'background_color': '#ffcbd1' 
-            }] 
-        },
-        {
-            'columns': ['C', 'D'],
-            'conditional_formats': [{
-                'filters': [{'condition': 'less_than_or_equal', 'value': 0}], 
-                'font_color': '#FFFFFF', 
-                'background_color': '#000000' 
-            }] 
-        },
-
-    ]
-    mito_backend = MitoBackend(column_definitions=column_definitions)
-
-    df_formats = mito_backend.steps_manager.curr_step.df_formats
-    
-    # Still generates valid, empty df_formats objects
-    assert df_formats[0]['columns'] == {}
-    assert df_formats[0]['headers'] == {}
-    assert df_formats[0]['rows']['even'] == {}
-    assert df_formats[0]['rows']['odd'] == {}
-    assert df_formats[0]['border'] == {}
-
-    # But now has conditional formats
-    assert len(df_formats[0]['conditional_formats']) == 2
-    assert df_formats[0]['conditional_formats'][0]['columnIDs'] == ['A', 'B']
-    assert df_formats[0]['conditional_formats'][0]['filters'] == [{'condition': 'greater', 'value': 5}]
-    assert df_formats[0]['conditional_formats'][0]['color'] == '#c30010'
-    assert df_formats[0]['conditional_formats'][0]['backgroundColor'] == '#ffcbd1'
-    assert df_formats[0]['conditional_formats'][0]['invalidFilterColumnIDs'] == []
-
-    assert df_formats[0]['conditional_formats'][1]['columnIDs'] == ['C', 'D']
-    assert df_formats[0]['conditional_formats'][1]['filters'] == [{'condition': 'less_than_or_equal', 'value': 0}]
-    assert df_formats[0]['conditional_formats'][1]['color'] == '#FFFFFF'
-    assert df_formats[0]['conditional_formats'][1]['backgroundColor'] == '#000000'
-    assert df_formats[0]['conditional_formats'][1]['invalidFilterColumnIDs'] == []
